Objects and links are userdata objects.
Objects have the properties id, name, archetype, position, facing,
and the methods Exists, HasMetaProperty, InheritsFrom, and SetName.
An object is created with the function lgs.link from either the object
id or the object name.

Links have the properties id, flavor, source, dest, and the methods
Get, GetData, SetData. A link is created with the function lgs.link
from the integer ID of the link.